ai-github-codebase-mentor/requirements.txt
streamlit
gitpython
sentence-transformers
transformers
torch
numpy
tqdm
python-dotenv
requests
pydantic
openai
google-generativeai
google-genai